refactor: move rusqlite impl into lib
