+++ src/db/config.rs
@@ -14,9 +14,3 @@ pub(super) fn database_path() -> anyhow::Result<PathBuf> {
     let base_dirs = base_directories()?;
     base_dirs.place_data_file("monotax.db")
 }
-
-pub fn connect() -> anyhow::Result<rusqlite::Connection> {
-    let db_path = database_path()?;
-    let conn = rusqlite::Connection::open(db_path)?;
-    Ok(conn)
-}
